Updated testing section, added readme(s)
The QA section now contains a readme to explain the usage of this project.
The testing section now contains a readme explaining how to use it to create a db/api/jenkin server in order to test those skills.
The testing section also contains a docker-compose.yaml file that can be used to create the servers with only one command.

The testing section will soon have a maven project contained within that will show automation framework practice.

Signed-off-by: H1ddenGames <skarim3000@gmail.com>