Updated testing folder with practice projects
The testing folder now contains a maven project that contains practice with the following:
1. API testing
2. Database Testing
3. Framework Testing
	3.1 BDT (Behaviour Driven Testing)
	3.2 BDTPOM (Behaviour Driven Testing + Page Object Model Hybrid)
	3.3 DDT (Data Driven Testing)
	3.4 KDT (Keyword Driven Testing)
	3.5 POM (Page Object Model)

The testing folder is within Testing/db-api-jenkins/jenkins/home because the project will be used within the Jenkins docker container to run automated headless tests through Jenkins.
Keeping project files within that folder means that the files are shared between the host machine and the docker container. This makes it possible to add the project as a local git project without having to create a github repository specifically for it.

Using this link as a proof of concept: https://stackoverflow.com/questions/10498554/jenkins-linking-to-my-local-git-repository

Signed-off-by: H1ddenGames <skarim3000@gmail.com>