Updated APITesting and DatabaseTesting
APITesting now has priorities set on testng annotations so that the tests can be ran over and over without any discrepancies between runs.
The tests have been ordered in such a way that if a request results in one or more resources being created or deleted, the next test with do the opposite of the last test.

DatabaseTesting has has it's layout finalized and the basic setup, teardown methods created of connecting/disconnecting from the database.

Signed-off-by: h1ddengames <skarim3000@gmail.com>